feat: update database

+++ src/main/java/org/example/backend/models/items/VehicleItem.java
@@ -2,6 +2,7 @@ package org.example.backend.models.items;
 
 import org.example.backend.models.Item;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class VehicleItem extends Item {
@@ -11,12 +12,12 @@ public class VehicleItem extends Item {
   protected double mileage; // số km đã đi
 
   public VehicleItem(
-      int id,
-      LocalDateTime createdAt,
-      String type,
-      String itemName,
-      String description,
-      double startingPrice) {
+          int id,
+          LocalDateTime createdAt,
+          String type,
+          String itemName,
+          String description,
+          BigDecimal startingPrice) {
     super(id, createdAt, type, itemName, description, startingPrice);
   }
